Restructures code
