+++ mnamer/core/utils.py
